Removed one-time depth buffer allocation from the driver. Removed depth buffer related values from the global (non context/drawable) structures and added them to the drawable. Initialize all visual configurations (16 and 32 bit).
Windowing system (dfb, miniglx) has to allocate the depth, too, now depending
on the values from the visual config that is to be used.

I'm now able to have two OpenGL windows (in one app for now) rendering
concurrently (not multi threaded yet, but easy to do) both using their
own depth buffer.

I'm sorry, but I had to disable the radeon driver until it's updated to
the recent changes. I also doubt that MiniGLX is currently working in this
branch.