Implement a locking scheme to protect texture object data.  This serves two purposes:
1) Prevent bound texture objects from statechanges by one context
while another context is referencing the data for rendering purposes.

2) Ensure that texture object state changed by one context provokes a
statechange notification (ie a bit set in ctx->NewState) in all other
sharing contexts.

This is a first pass which uses a single mutex for all texture objects
in a share group.  A future step would be to use a mutex per texobj.