Begin to deal with full batch buffers.
Move the state differencer out of the regular state mechanism.
Split state handling into two steps - first, update software state
and calculate what we think hardware state should be.  This is the
old state mechanism.  The new, second phase which is delayed until
the very point at which triangles or other primitives are emitted
to the hardware is to perform the state difference step and emit
the state to hardware.

This is done with the information about the required batchbuffer
space for the primitive emit, so that there is a guarentee that
the state emit and primitive can both fit in the current batch
buffer.  If the batchbuffer wraps, the full state is emitted to cope
with potential lost contexts.

Also, split the vertex-buffer handling out of intel_idx_render.c and
make it a standalone facility.  Move the old intel_tris.c/intel_render.c
code to use this facility to allocate vb space for their primitives and have
them issue indirect primitives rather than inline ones.

Reinstate the render stage in the pipeline above the indexed render stage,
as it should be slightly more efficient when it is available.