[965] Convert the remaining parts of the driver to dri_emit_reloc().
This is ugly for the current caching scheme, as we stick in stub values
for the offsets of referenced buffers, resulting in excessive cache misses.

However, the cache needs to be rewritten anyway due to being a giant
ringbuffered buffer-of-buffers instead of just buffer objects per cached state
buffer. At that point, we can create better keys for searching for cache hits.

It also results in a bit more batchbuffered state being uploaded when buffer
offsets stay the same, since we don't know that they'll be the same until
batch submit.