[965] Move relocation emits to a separate hook and remove bogus relocations.
The relocation emits have to be separate as they need to be called per
state validation (batchbuffer submit), not just when a new state buffer is
prepared.  Additionally, most of the state doesn't need relocations, as
all of the state is in one of two buffer objects (BRW_SS_POOL and
BRW_GS_POOL), and the offsets programmed in state buffers are relative to
the base of one of those pools (buffer objects), so most of the relocation
emits I had added were bogus.