st/vbo: clone vbo module and rename structs to st_vbo
Want to acheive a couple of things here:
1) Build a better immediate vertex handler
2) Build gallium vertex buffer/element state directly from st_vbo module
3) Avoid breaking all the DRI drivers in the process

The vbo code is pretty fragile and even small changes there seem to
break DRI drivers left & right.  In part that's due to the lack of a
strong interface between the shared code and the drivers, and also to
the fairly ill-defined interface between core mesa and the VBO module.

Hopefully this branch will help figure out a better interface between
mesa and the vertex code, and will be able to do so without breaking
other drivers in the process.

If this branch turns out well, it may be interesting to back-port some
of the changes here to the original VBO module.