gallium: clarify sw driver api
Seem to be gravitating to a design where a slightly trimmed
llvmpipe winsys struct is passed to software drivers at create_screen().

This keeps the existing flow in the software drivers, but differs
from the way display targets and textures are associated with each
other in the DRI2 world.

For the moment at least it seems like we will be creating display
target textures in very different ways in software vs. hardware drivers.