GLSL2->LLVM conversion code and "glslang" compiler
It should support everything except texture and derivative operations.
It seems to work.

This is right now only a proof of concept, and there is no graphics
pipeline you can use the result in.

Comments, enhancements, etc. welcome.

This can be merged into master if the GLSL2 maintainers want that.