softpipe: rework blending to work properly
This changes makes softpipe clamp blend source, destination input,
source factor and destination factor if the target is fixed point.

Also, it makes it support an unclamped blend color properly.

Note this is NOT related to vertex/fragment clamping and is
independent of these controls, and only dependent on the type of
_each_ color buffer.

This is supposed to be the behavior specified by OpenGL 4.1 and
ARB_color_buffer_float.

Not sure about other APIs.