llvmpipe: switch from byte-filled tiles to float-filled tiles (v2)
Changes in v2:
- don't try to use SSE3 swizzles, fixes x86-64 build

This is probably undesirable in this form, since it will
quadruplicate the memory bandwidth and cache occupation for all
resources. It might however reduce the number of instructions.

(It also quadruplicates RAM usage, but we are fine with this, since
system RAM is typically four times the size of VRAM)

Making a byte-based llvmpipe and a float-based one coexist should
be the appropriate solution, but is not done in this patch currently.