llvmpipe: rework blending to work properly
Blending has likely been broken from the move to float tiles, since
values are no longer automatically clamped.

Also, we need to rework it to properly support blending for both
fixed and floating point formats.

Note that this approach is probably suboptimal, and we should instead
always keep tiles with whatever data type the resource has, and blend
in the resource data type.

Blending in the resource data type gives the correct clamping semantics
for free, and should be faster.