mesa: add MESA_ROBUST_ACCESS env var to enable ARB_robustness
We use the robust access bit to choose whether to apply some loop
optimizations in the GLSL compiler, so it's a good idea to be able
to test this, and it's useful in general to be able to force robustness
from the command line.

In the future, GLX_ARB_create_context_robustness will also allow to turn
this on programmatically in the standard way.