glsl2: Fix linear portion of index search for writes
If the array access is a read, read the first element of this
subregion unconditionally.  The remaining tests will possibly
overwrite this value with one of the other array elements.

This optimization cannot be done for writes because it will cause the
first element of the subregion to be written possibly *in addition* to
one of the other elements.